test(nexus): move the rate-limit identifier test into the nexus tree (#2027)
The root `pytest.ini` puts only core's `src` on the path, so a root-suite test
importing `nexus` resolved the pip-installed package rather than the tree under
test. In a worktree that is a DIFFERENT checkout, so the test reported on
source it was not exercising -- it failed against fixed code because it was
reading the unfixed install.

Relocated beside the middleware it covers, where the package's own conftest
puts the correct `src` first. Verified it resolves the worktree source and
reds when middleware.py is reverted.

Refs #2027
